backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
openai-agents-sdk==0.1.0
official-mcp-sdk==0.1.0
asyncpg==0.29.0
better-auth==0.0.1
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.0